+++ sdk/resourcehealth/azure-mgmt-resourcehealth/azure/mgmt/resourcehealth/models/_models_py3.py
@@ -1,5 +1,5 @@
+# pylint: disable=line-too-long,useless-suppression,too-many-lines
 # coding=utf-8
-# pylint: disable=too-many-lines
 # --------------------------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
@@ -7,21 +7,15 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
+from collections.abc import MutableMapping
 import datetime
-import sys
 from typing import Any, List, Optional, TYPE_CHECKING, Union
 
-from .. import _serialization
-
-if sys.version_info >= (3, 9):
-    from collections.abc import MutableMapping
-else:
-    from typing import MutableMapping  # type: ignore  # pylint: disable=ungrouped-imports
+from .._utils import serialization as _serialization
 
 if TYPE_CHECKING:
-    # pylint: disable=unused-import,ungrouped-imports
     from .. import models as _models
-JSON = MutableMapping[str, Any]  # pylint: disable=unsubscriptable-object
+JSON = MutableMapping[str, Any]
 
 
 class AvailabilityStatus(_serialization.Model):
@@ -113,7 +107,7 @@ class AvailabilityStatusListResult(_serialization.Model):
         self.next_link = next_link
 
 
-class AvailabilityStatusProperties(_serialization.Model):  # pylint: disable=too-many-instance-attributes
+class AvailabilityStatusProperties(_serialization.Model):
     """Properties of availability state.
 
     :ivar availability_state: Availability status of the resource. When it is null, this
@@ -430,7 +424,7 @@ class Resource(_serialization.Model):
     Variables are only populated by the server, and will be ignored when sending a request.
 
     :ivar id: Fully qualified resource ID for the resource. Ex -
-     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.  # pylint: disable=line-too-long
+     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.
     :vartype id: str
     :ivar name: The name of the resource.
     :vartype name: str
@@ -459,10 +453,10 @@ class Resource(_serialization.Model):
     def __init__(self, **kwargs: Any) -> None:
         """ """
         super().__init__(**kwargs)
-        self.id = None
-        self.name = None
-        self.type = None
-        self.system_data = None
+        self.id: Optional[str] = None
+        self.name: Optional[str] = None
+        self.type: Optional[str] = None
+        self.system_data: Optional["_models.SystemData"] = None
 
 
 class ProxyResource(Resource):
@@ -472,7 +466,7 @@ class ProxyResource(Resource):
     Variables are only populated by the server, and will be ignored when sending a request.
 
     :ivar id: Fully qualified resource ID for the resource. Ex -
-     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.  # pylint: disable=line-too-long
+     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.
     :vartype id: str
     :ivar name: The name of the resource.
     :vartype name: str
@@ -491,7 +485,7 @@ class EmergingIssuesGetResult(ProxyResource):
     Variables are only populated by the server, and will be ignored when sending a request.
 
     :ivar id: Fully qualified resource ID for the resource. Ex -
-     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.  # pylint: disable=line-too-long
+     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.
     :vartype id: str
     :ivar name: The name of the resource.
     :vartype name: str
@@ -576,18 +570,18 @@ class ErrorResponse(_serialization.Model):
     def __init__(self, **kwargs: Any) -> None:
         """ """
         super().__init__(**kwargs)
-        self.code = None
-        self.message = None
-        self.details = None
+        self.code: Optional[str] = None
+        self.message: Optional[str] = None
+        self.details: Optional[str] = None
 
 
-class Event(ProxyResource):  # pylint: disable=too-many-instance-attributes
+class Event(ProxyResource):
     """Service health event.
 
     Variables are only populated by the server, and will be ignored when sending a request.
 
     :ivar id: Fully qualified resource ID for the resource. Ex -
-     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.  # pylint: disable=line-too-long
+     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.
     :vartype id: str
     :ivar name: The name of the resource.
     :vartype name: str
@@ -600,16 +594,14 @@ class Event(ProxyResource):  # pylint: disable=too-many-instance-attributes
     :ivar event_type: Type of event. Known values are: "ServiceIssue", "PlannedMaintenance",
      "HealthAdvisory", "RCA", "EmergingIssues", and "SecurityAdvisory".
     :vartype event_type: str or ~azure.mgmt.resourcehealth.models.EventTypeValues
-    :ivar event_sub_type: Sub type of the event. Currently used to determine retirement
-     communications for health advisory events. "Retirement"
-    :vartype event_sub_type: str or ~azure.mgmt.resourcehealth.models.EventSubTypeValues
     :ivar event_source: Source of event. Known values are: "ResourceHealth" and "ServiceHealth".
     :vartype event_source: str or ~azure.mgmt.resourcehealth.models.EventSourceValues
     :ivar status: Current status of event. Known values are: "Active" and "Resolved".
     :vartype status: str or ~azure.mgmt.resourcehealth.models.EventStatusValues
     :ivar title: Title text of event.
     :vartype title: str
-    :ivar summary: Summary text of event.
+    :ivar summary: Summary text of event. Use fetchEventDetails endpoint to get summary of
+     sensitive events.
     :vartype summary: str
     :ivar header: Header text of event.
     :vartype header: str
@@ -618,6 +610,10 @@ class Event(ProxyResource):  # pylint: disable=too-many-instance-attributes
     :ivar event_level: Level of event. Known values are: "Critical", "Error", "Warning", and
      "Informational".
     :vartype event_level: str or ~azure.mgmt.resourcehealth.models.EventLevelValues
+    :ivar is_event_sensitive: If true the event may contains sensitive data. Use the post
+     events/{trackingId}/fetchEventDetails endpoint to fetch sensitive data see
+     https://learn.microsoft.com/en-us/azure/service-health/security-advisories-elevated-access.
+    :vartype is_event_sensitive: bool
     :ivar external_incident_id: The id of the Incident.
     :vartype external_incident_id: str
     :ivar reason: The reason for the Incident.
@@ -644,7 +640,8 @@ class Event(ProxyResource):  # pylint: disable=too-many-instance-attributes
      this event.
     :vartype enable_microsoft_support: bool
     :ivar description: Contains the communication message for the event, that could include
-     summary, root cause and other details.
+     summary, root cause and other details. Use fetchEventDetails endpoint to get description of
+     sensitive events.
     :vartype description: str
     :ivar platform_initiated: Is true if the event is platform initiated.
     :vartype platform_initiated: bool
@@ -668,13 +665,21 @@ class Event(ProxyResource):  # pylint: disable=too-many-instance-attributes
     :vartype duration: int
     :ivar impact_type: The type of the impact.
     :vartype impact_type: str
-    :ivar maintenance_id: Unique identifier for planned maintenance event.
-    :vartype maintenance_id: str
-    :ivar maintenance_type: The type of planned maintenance event.
-    :vartype maintenance_type: str
-    :ivar arg_query: Azure Resource Graph query to fetch the affected resources from their existing
-     Azure Resource Graph locations.
-    :vartype arg_query: str
+    :ivar event_tags: A list of metadata tags associated with the event. Possible values include:
+      -Action Recommended: Action may be required by you to avoid possible disruptions or mitigate
+     risks for your services. It is recommended to evaluate these actions and the potential impact
+     on your services.
+
+
+     * False Positive: After investigation, we've determined your service is healthy and service
+     issues did not impact your services as originally communicated.
+     * Preliminary PIR: For our largest, most impactful service issues a Preliminary Post Incident
+     Review (PIR) is published generally within 72 hours of mitigation, to summarize what we have
+     learned so far from the still-in-progress investigation.
+     * Final PIR: For service issues, a Final Post Incident Review (PIR) may be published to provide
+     additional details or learnings. Sometimes this requires us to complete an internal
+     retrospective, generally within 14 days of mitigation.
+    :vartype event_tags: list[str]
     """
 
     _validation = {
@@ -690,7 +695,6 @@ class Event(ProxyResource):  # pylint: disable=too-many-instance-attributes
         "type": {"key": "type", "type": "str"},
         "system_data": {"key": "systemData", "type": "SystemData"},
         "event_type": {"key": "properties.eventType", "type": "str"},
-        "event_sub_type": {"key": "properties.eventSubType", "type": "str"},
         "event_source": {"key": "properties.eventSource", "type": "str"},
         "status": {"key": "properties.status", "type": "str"},
         "title": {"key": "properties.title", "type": "str"},
@@ -698,6 +702,7 @@ class Event(ProxyResource):  # pylint: disable=too-many-instance-attributes
         "header": {"key": "properties.header", "type": "str"},
         "level": {"key": "properties.level", "type": "str"},
         "event_level": {"key": "properties.eventLevel", "type": "str"},
+        "is_event_sensitive": {"key": "properties.isEventSensitive", "type": "bool"},
         "external_incident_id": {"key": "properties.externalIncidentId", "type": "str"},
         "reason": {"key": "properties.reason", "type": "str"},
         "article": {"key": "properties.article", "type": "EventPropertiesArticle"},
@@ -721,16 +726,13 @@ class Event(ProxyResource):  # pylint: disable=too-many-instance-attributes
         },
         "duration": {"key": "properties.duration", "type": "int"},
         "impact_type": {"key": "properties.impactType", "type": "str"},
-        "maintenance_id": {"key": "properties.maintenanceId", "type": "str"},
-        "maintenance_type": {"key": "properties.maintenanceType", "type": "str"},
-        "arg_query": {"key": "properties.argQuery", "type": "str"},
+        "event_tags": {"key": "properties.eventTags", "type": "[str]"},
     }
 
     def __init__(  # pylint: disable=too-many-locals
         self,
         *,
         event_type: Optional[Union[str, "_models.EventTypeValues"]] = None,
-        event_sub_type: Optional[Union[str, "_models.EventSubTypeValues"]] = None,
         event_source: Optional[Union[str, "_models.EventSourceValues"]] = None,
         status: Optional[Union[str, "_models.EventStatusValues"]] = None,
         title: Optional[str] = None,
@@ -738,6 +740,7 @@ class Event(ProxyResource):  # pylint: disable=too-many-instance-attributes
         header: Optional[str] = None,
         level: Optional[Union[str, "_models.LevelValues"]] = None,
         event_level: Optional[Union[str, "_models.EventLevelValues"]] = None,
+        is_event_sensitive: Optional[bool] = None,
         external_incident_id: Optional[str] = None,
         reason: Optional[str] = None,
         article: Optional["_models.EventPropertiesArticle"] = None,
@@ -758,25 +761,21 @@ class Event(ProxyResource):  # pylint: disable=too-many-instance-attributes
         additional_information: Optional["_models.EventPropertiesAdditionalInformation"] = None,
         duration: Optional[int] = None,
         impact_type: Optional[str] = None,
-        maintenance_id: Optional[str] = None,
-        maintenance_type: Optional[str] = None,
-        arg_query: Optional[str] = None,
+        event_tags: Optional[List[str]] = None,
         **kwargs: Any
     ) -> None:
         """
         :keyword event_type: Type of event. Known values are: "ServiceIssue", "PlannedMaintenance",
          "HealthAdvisory", "RCA", "EmergingIssues", and "SecurityAdvisory".
         :paramtype event_type: str or ~azure.mgmt.resourcehealth.models.EventTypeValues
-        :keyword event_sub_type: Sub type of the event. Currently used to determine retirement
-         communications for health advisory events. "Retirement"
-        :paramtype event_sub_type: str or ~azure.mgmt.resourcehealth.models.EventSubTypeValues
         :keyword event_source: Source of event. Known values are: "ResourceHealth" and "ServiceHealth".
         :paramtype event_source: str or ~azure.mgmt.resourcehealth.models.EventSourceValues
         :keyword status: Current status of event. Known values are: "Active" and "Resolved".
         :paramtype status: str or ~azure.mgmt.resourcehealth.models.EventStatusValues
         :keyword title: Title text of event.
         :paramtype title: str
-        :keyword summary: Summary text of event.
+        :keyword summary: Summary text of event. Use fetchEventDetails endpoint to get summary of
+         sensitive events.
         :paramtype summary: str
         :keyword header: Header text of event.
         :paramtype header: str
@@ -785,6 +784,10 @@ class Event(ProxyResource):  # pylint: disable=too-many-instance-attributes
         :keyword event_level: Level of event. Known values are: "Critical", "Error", "Warning", and
          "Informational".
         :paramtype event_level: str or ~azure.mgmt.resourcehealth.models.EventLevelValues
+        :keyword is_event_sensitive: If true the event may contains sensitive data. Use the post
+         events/{trackingId}/fetchEventDetails endpoint to fetch sensitive data see
+         https://learn.microsoft.com/en-us/azure/service-health/security-advisories-elevated-access.
+        :paramtype is_event_sensitive: bool
         :keyword external_incident_id: The id of the Incident.
         :paramtype external_incident_id: str
         :keyword reason: The reason for the Incident.
@@ -812,7 +815,8 @@ class Event(ProxyResource):  # pylint: disable=too-many-instance-attributes
          this event.
         :paramtype enable_microsoft_support: bool
         :keyword description: Contains the communication message for the event, that could include
-         summary, root cause and other details.
+         summary, root cause and other details. Use fetchEventDetails endpoint to get description of
+         sensitive events.
         :paramtype description: str
         :keyword platform_initiated: Is true if the event is platform initiated.
         :paramtype platform_initiated: bool
@@ -836,17 +840,25 @@ class Event(ProxyResource):  # pylint: disable=too-many-instance-attributes
         :paramtype duration: int
         :keyword impact_type: The type of the impact.
         :paramtype impact_type: str
-        :keyword maintenance_id: Unique identifier for planned maintenance event.
-        :paramtype maintenance_id: str
-        :keyword maintenance_type: The type of planned maintenance event.
-        :paramtype maintenance_type: str
-        :keyword arg_query: Azure Resource Graph query to fetch the affected resources from their
-         existing Azure Resource Graph locations.
-        :paramtype arg_query: str
+        :keyword event_tags: A list of metadata tags associated with the event. Possible values
+         include:
+          -Action Recommended: Action may be required by you to avoid possible disruptions or mitigate
+         risks for your services. It is recommended to evaluate these actions and the potential impact
+         on your services.
+
+
+         * False Positive: After investigation, we've determined your service is healthy and service
+         issues did not impact your services as originally communicated.
+         * Preliminary PIR: For our largest, most impactful service issues a Preliminary Post Incident
+         Review (PIR) is published generally within 72 hours of mitigation, to summarize what we have
+         learned so far from the still-in-progress investigation.
+         * Final PIR: For service issues, a Final Post Incident Review (PIR) may be published to provide
+         additional details or learnings. Sometimes this requires us to complete an internal
+         retrospective, generally within 14 days of mitigation.
+        :paramtype event_tags: list[str]
         """
         super().__init__(**kwargs)
         self.event_type = event_type
-        self.event_sub_type = event_sub_type
         self.event_source = event_source
         self.status = status
         self.title = title
@@ -854,6 +866,7 @@ class Event(ProxyResource):  # pylint: disable=too-many-instance-attributes
         self.header = header
         self.level = level
         self.event_level = event_level
+        self.is_event_sensitive = is_event_sensitive
         self.external_incident_id = external_incident_id
         self.reason = reason
         self.article = article
@@ -874,18 +887,16 @@ class Event(ProxyResource):  # pylint: disable=too-many-instance-attributes
         self.additional_information = additional_information
         self.duration = duration
         self.impact_type = impact_type
-        self.maintenance_id = maintenance_id
-        self.maintenance_type = maintenance_type
-        self.arg_query = arg_query
+        self.event_tags = event_tags
 
 
-class EventImpactedResource(ProxyResource):  # pylint: disable=too-many-instance-attributes
+class EventImpactedResource(ProxyResource):
     """Impacted resource for an event.
 
     Variables are only populated by the server, and will be ignored when sending a request.
 
     :ivar id: Fully qualified resource ID for the resource. Ex -
-     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.  # pylint: disable=line-too-long
+     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.
     :vartype id: str
     :ivar name: The name of the resource.
     :vartype name: str
@@ -901,16 +912,6 @@ class EventImpactedResource(ProxyResource):  # pylint: disable=too-many-instance
     :vartype target_resource_id: str
     :ivar target_region: Impacted resource region name.
     :vartype target_region: str
-    :ivar resource_name: Resource name of the impacted resource.
-    :vartype resource_name: str
-    :ivar resource_group: Resource group name of the impacted resource.
-    :vartype resource_group: str
-    :ivar status: Status of the impacted resource.
-    :vartype status: str
-    :ivar maintenance_start_time: Start time of maintenance for the impacted resource.
-    :vartype maintenance_start_time: str
-    :ivar maintenance_end_time: End time of maintenance for the impacted resource.
-    :vartype maintenance_end_time: str
     :ivar info: Additional information.
     :vartype info: list[~azure.mgmt.resourcehealth.models.KeyValueItem]
     """
@@ -923,11 +924,6 @@ class EventImpactedResource(ProxyResource):  # pylint: disable=too-many-instance
         "target_resource_type": {"readonly": True},
         "target_resource_id": {"readonly": True},
         "target_region": {"readonly": True},
-        "resource_name": {"readonly": True},
-        "resource_group": {"readonly": True},
-        "status": {"readonly": True},
-        "maintenance_start_time": {"readonly": True},
-        "maintenance_end_time": {"readonly": True},
     }
 
     _attribute_map = {
@@ -938,11 +934,6 @@ class EventImpactedResource(ProxyResource):  # pylint: disable=too-many-instance
         "target_resource_type": {"key": "properties.targetResourceType", "type": "str"},
         "target_resource_id": {"key": "properties.targetResourceId", "type": "str"},
         "target_region": {"key": "properties.targetRegion", "type": "str"},
-        "resource_name": {"key": "properties.resourceName", "type": "str"},
-        "resource_group": {"key": "properties.resourceGroup", "type": "str"},
-        "status": {"key": "properties.status", "type": "str"},
-        "maintenance_start_time": {"key": "properties.maintenanceStartTime", "type": "str"},
-        "maintenance_end_time": {"key": "properties.maintenanceEndTime", "type": "str"},
         "info": {"key": "properties.info", "type": "[KeyValueItem]"},
     }
 
@@ -952,14 +943,9 @@ class EventImpactedResource(ProxyResource):  # pylint: disable=too-many-instance
         :paramtype info: list[~azure.mgmt.resourcehealth.models.KeyValueItem]
         """
         super().__init__(**kwargs)
-        self.target_resource_type = None
-        self.target_resource_id = None
-        self.target_region = None
-        self.resource_name = None
-        self.resource_group = None
-        self.status = None
-        self.maintenance_start_time = None
-        self.maintenance_end_time = None
+        self.target_resource_type: Optional[str] = None
+        self.target_resource_id: Optional[str] = None
+        self.target_region: Optional[str] = None
         self.info = info
 
 
@@ -1202,12 +1188,16 @@ class Impact(_serialization.Model):
 
     :ivar impacted_service: Impacted service name.
     :vartype impacted_service: str
+    :ivar impacted_service_guid: Impacted service guid. This is the permanent identifier for the
+     impacted service.
+    :vartype impacted_service_guid: str
     :ivar impacted_regions: List regions impacted by the service health event.
     :vartype impacted_regions: list[~azure.mgmt.resourcehealth.models.ImpactedServiceRegion]
     """
 
     _attribute_map = {
         "impacted_service": {"key": "impactedService", "type": "str"},
+        "impacted_service_guid": {"key": "impactedServiceGuid", "type": "str"},
         "impacted_regions": {"key": "impactedRegions", "type": "[ImpactedServiceRegion]"},
     }
 
@@ -1215,17 +1205,22 @@ class Impact(_serialization.Model):
         self,
         *,
         impacted_service: Optional[str] = None,
+        impacted_service_guid: Optional[str] = None,
         impacted_regions: Optional[List["_models.ImpactedServiceRegion"]] = None,
         **kwargs: Any
     ) -> None:
         """
         :keyword impacted_service: Impacted service name.
         :paramtype impacted_service: str
+        :keyword impacted_service_guid: Impacted service guid. This is the permanent identifier for the
+         impacted service.
+        :paramtype impacted_service_guid: str
         :keyword impacted_regions: List regions impacted by the service health event.
         :paramtype impacted_regions: list[~azure.mgmt.resourcehealth.models.ImpactedServiceRegion]
         """
         super().__init__(**kwargs)
         self.impacted_service = impacted_service
+        self.impacted_service_guid = impacted_service_guid
         self.impacted_regions = impacted_regions
 
 
@@ -1267,7 +1262,7 @@ class ImpactedResourceStatus(ProxyResource):
     Variables are only populated by the server, and will be ignored when sending a request.
 
     :ivar id: Fully qualified resource ID for the resource. Ex -
-     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.  # pylint: disable=line-too-long
+     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.
     :vartype id: str
     :ivar name: The name of the resource.
     :vartype name: str
@@ -1358,7 +1353,8 @@ class ImpactedServiceRegion(_serialization.Model):
     :ivar last_update_time: It provides the Timestamp for when the last update for the service
      health event.
     :vartype last_update_time: ~datetime.datetime
-    :ivar updates: List of updates for given service health event.
+    :ivar updates: List of updates for given service health event.  Use fetchEventDetails endpoint
+     to get updates of sensitive events.
     :vartype updates: list[~azure.mgmt.resourcehealth.models.Update]
     """
 
@@ -1395,7 +1391,8 @@ class ImpactedServiceRegion(_serialization.Model):
         :keyword last_update_time: It provides the Timestamp for when the last update for the service
          health event.
         :paramtype last_update_time: ~datetime.datetime
-        :keyword updates: List of updates for given service health event.
+        :keyword updates: List of updates for given service health event.  Use fetchEventDetails
+         endpoint to get updates of sensitive events.
         :paramtype updates: list[~azure.mgmt.resourcehealth.models.Update]
         """
         super().__init__(**kwargs)
@@ -1431,8 +1428,8 @@ class KeyValueItem(_serialization.Model):
     def __init__(self, **kwargs: Any) -> None:
         """ """
         super().__init__(**kwargs)
-        self.key = None
-        self.value = None
+        self.key: Optional[str] = None
+        self.value: Optional[str] = None
 
 
 class Link(_serialization.Model):
@@ -1526,7 +1523,7 @@ class MetadataEntity(ProxyResource):
     Variables are only populated by the server, and will be ignored when sending a request.
 
     :ivar id: Fully qualified resource ID for the resource. Ex -
-     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.  # pylint: disable=line-too-long
+     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.
     :vartype id: str
     :ivar name: The name of the resource.
     :vartype name: str
@@ -1622,8 +1619,12 @@ class MetadataEntityListResult(_serialization.Model):
 class MetadataSupportedValueDetail(_serialization.Model):
     """The metadata supported value detail.
 
-    :ivar id: The id.
+    :ivar id: The id of the metadata value.
     :vartype id: str
+    :ivar previous_id: The previous value of the id field incase the data has changed.
+    :vartype previous_id: str
+    :ivar service_guid: The permanent guid for the service. Used when the id is a service name.
+    :vartype service_guid: str
     :ivar display_name: The display name.
     :vartype display_name: str
     :ivar resource_types: The list of associated resource types.
@@ -1632,6 +1633,8 @@ class MetadataSupportedValueDetail(_serialization.Model):
 
     _attribute_map = {
         "id": {"key": "id", "type": "str"},
+        "previous_id": {"key": "previousId", "type": "str"},
+        "service_guid": {"key": "serviceGuid", "type": "str"},
         "display_name": {"key": "displayName", "type": "str"},
         "resource_types": {"key": "resourceTypes", "type": "[str]"},
     }
@@ -1640,13 +1643,19 @@ class MetadataSupportedValueDetail(_serialization.Model):
         self,
         *,
         id: Optional[str] = None,  # pylint: disable=redefined-builtin
+        previous_id: Optional[str] = None,
+        service_guid: Optional[str] = None,
         display_name: Optional[str] = None,
         resource_types: Optional[List[str]] = None,
         **kwargs: Any
     ) -> None:
         """
-        :keyword id: The id.
+        :keyword id: The id of the metadata value.
         :paramtype id: str
+        :keyword previous_id: The previous value of the id field incase the data has changed.
+        :paramtype previous_id: str
+        :keyword service_guid: The permanent guid for the service. Used when the id is a service name.
+        :paramtype service_guid: str
         :keyword display_name: The display name.
         :paramtype display_name: str
         :keyword resource_types: The list of associated resource types.
@@ -1654,6 +1663,8 @@ class MetadataSupportedValueDetail(_serialization.Model):
         """
         super().__init__(**kwargs)
         self.id = id
+        self.previous_id = previous_id
+        self.service_guid = service_guid
         self.display_name = display_name
         self.resource_types = resource_types
 
@@ -2138,15 +2149,36 @@ class Update(_serialization.Model):
     :ivar update_date_time: It provides the Timestamp for the given update for the service health
      event.
     :vartype update_date_time: ~datetime.datetime
+    :ivar event_tags: A list of metadata tags associated with the event. Possible values include:
+      -Action Recommended: Action may be required by you to avoid possible disruptions or mitigate
+     risks for your services. It is recommended to evaluate these actions and the potential impact
+     on your services.
+
+
+     * False Positive: After investigation, we've determined your service is healthy and service
+     issues did not impact your services as originally communicated.
+     * Preliminary PIR: For our largest, most impactful service issues a Preliminary Post Incident
+     Review (PIR) is published generally within 72 hours of mitigation, to summarize what we have
+     learned so far from the still-in-progress investigation.
+     * Final PIR: For service issues, a Final Post Incident Review (PIR) may be published to provide
+     additional details or learnings. Sometimes this requires us to complete an internal
+     retrospective, generally within 14 days of mitigation.
+    :vartype event_tags: list[str]
     """
 
     _attribute_map = {
         "summary": {"key": "summary", "type": "str"},
         "update_date_time": {"key": "updateDateTime", "type": "iso-8601"},
+        "event_tags": {"key": "eventTags", "type": "[str]"},
     }
 
     def __init__(
-        self, *, summary: Optional[str] = None, update_date_time: Optional[datetime.datetime] = None, **kwargs: Any
+        self,
+        *,
+        summary: Optional[str] = None,
+        update_date_time: Optional[datetime.datetime] = None,
+        event_tags: Optional[List[str]] = None,
+        **kwargs: Any
     ) -> None:
         """
         :keyword summary: Summary text for the given update for the service health event.
@@ -2154,7 +2186,24 @@ class Update(_serialization.Model):
         :keyword update_date_time: It provides the Timestamp for the given update for the service
          health event.
         :paramtype update_date_time: ~datetime.datetime
+        :keyword event_tags: A list of metadata tags associated with the event. Possible values
+         include:
+          -Action Recommended: Action may be required by you to avoid possible disruptions or mitigate
+         risks for your services. It is recommended to evaluate these actions and the potential impact
+         on your services.
+
+
+         * False Positive: After investigation, we've determined your service is healthy and service
+         issues did not impact your services as originally communicated.
+         * Preliminary PIR: For our largest, most impactful service issues a Preliminary Post Incident
+         Review (PIR) is published generally within 72 hours of mitigation, to summarize what we have
+         learned so far from the still-in-progress investigation.
+         * Final PIR: For service issues, a Final Post Incident Review (PIR) may be published to provide
+         additional details or learnings. Sometimes this requires us to complete an internal
+         retrospective, generally within 14 days of mitigation.
+        :paramtype event_tags: list[str]
         """
         super().__init__(**kwargs)
         self.summary = summary
         self.update_date_time = update_date_time
+        self.event_tags = event_tags